+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.9 branch
 ### 1.9.2
+* `url_to_absolute()` had been deprecated and has been removed;
 * `slug()` global function had been deprecated and has been removed.
 
 ### 1.9.1-RC2